requirements.txt
clickhouse-connect==0.7.15
pandas
sqlalchemy
python-dotenv
